Rename and repackage callback to YgoprodeckResultCallback
